Unit test for garbage collection|
